requirements.txt
tensorflow
numpy
pandas
matplotlib
pillow
streamlit
scikit-learn
mlflow
jupyter
notebook
ipykernel